+++ blocklets/ai-studio/src/components/file-editor/output/type.tsx
@@ -1,19 +1,10 @@
-import AigneLogo from '@app/icons/aigne-logo';
-import AigneLogoOutput from '@app/icons/aigne-logo-output';
 import { OutputVariable, OutputVariableYjs, RuntimeOutputVariable } from '@blocklet/ai-runtime/types';
 import { Icon } from '@iconify-icon/react';
-import FormsIcon from '@iconify-icons/tabler/forms';
-import LayoutIcon from '@iconify-icons/tabler/layout';
 import LinkIcon from '@iconify-icons/tabler/link';
-import MessageIcon from '@iconify-icons/tabler/message';
-import MessageCircleQuestionIcon from '@iconify-icons/tabler/message-circle-question';
 import MessageQuestionIcon from '@iconify-icons/tabler/message-question';
 import MessagesIcon from '@iconify-icons/tabler/messages';
 import PhotoIcon from '@iconify-icons/tabler/photo';
-import ShareIcon from '@iconify-icons/tabler/share';
-import UserIcon from '@iconify-icons/tabler/user';
 import WritingIcon from '@iconify-icons/tabler/writing';
-import { Box } from '@mui/material';
 import { ReactNode } from 'react';
 
 export const runtimeOutputVariables: {
@@ -53,63 +44,10 @@ export const runtimeOutputVariables: {
         name: RuntimeOutputVariable.referenceLinks,
       },
       {
-        icon: <Box component={AigneLogo} fontSize={14} />,
-        title: 'Children',
-        i18nKey: 'children',
-        name: RuntimeOutputVariable.children,
-      },
-      {
         icon: <Icon icon={MessagesIcon} />,
         title: 'LLM Response Stream',
         i18nKey: 'llmResponseStream',
         name: RuntimeOutputVariable.llmResponseStream,
-      },
-    ],
-  },
-  {
-    group: 'appearance',
-    outputs: [
-      {
-        icon: <Icon icon={UserIcon} />,
-        title: 'Profile',
-        i18nKey: 'profile',
-        name: RuntimeOutputVariable.profile,
-      },
-      {
-        icon: <Icon icon={LayoutIcon} />,
-        title: 'Appearance Page',
-        i18nKey: 'appearancePage',
-        name: RuntimeOutputVariable.appearancePage,
-      },
-      {
-        icon: <Icon icon={FormsIcon} />,
-        title: 'Appearance Input',
-        i18nKey: 'appearanceInput',
-        name: RuntimeOutputVariable.appearanceInput,
-      },
-      {
-        icon: <Box component={AigneLogoOutput} fontSize={14} />,
-        title: 'Appearance Output',
-        i18nKey: 'appearanceOutput',
-        name: RuntimeOutputVariable.appearanceOutput,
-      },
-      {
-        icon: <Icon icon={ShareIcon} />,
-        title: 'Share',
-        i18nKey: 'share',
-        name: RuntimeOutputVariable.share,
-      },
-      {
-        icon: <Icon icon={MessageCircleQuestionIcon} />,
-        title: 'Opening Questions',
-        i18nKey: 'openingQuestions',
-        name: RuntimeOutputVariable.openingQuestions,
-      },
-      {
-        icon: <Icon icon={MessageIcon} />,
-        title: 'Opening Message',
-        i18nKey: 'openingMessage',
-        name: RuntimeOutputVariable.openingMessage,
       },
     ],
   },