+++ blocklets/ai-studio/src/pages/project/project-page.tsx
@@ -41,6 +41,7 @@ import DiscussView from './discuss-view';
 import FileTree, { ImperativeFileTree } from './file-tree';
 import DeveloperTools from './icons/developer-tools';
 import Empty from './icons/empty';
+import PreviewView from './preview-view';
 import { useProjectState } from './state';
 import { newDefaultPrompt } from './template';
 import TestView from './test-view';
@@ -90,7 +91,7 @@ function ProjectPageView() {
   const navigate = useNavigate();
   useTitle(project?.name || 'AI Studio');
 
-  const [currentTab, setCurrentTab] = useLocalStorageState(CURRENT_TAB(projectId), { defaultValue: 'debug' });
+  const [currentTab, setCurrentTab] = useLocalStorageState(CURRENT_TAB(projectId), { defaultValue: 'preview' });
 
   const [previousFilePath, setPreviousFilePath] = useLocalStorageState<{ [key: string]: string } | undefined>(
     PREVIOUS_FILE_PATH(projectId)
@@ -287,6 +288,7 @@ function ProjectPageView() {
                     },
                   },
                 }}>
+                <Tab value="preview" label={t('preview')} data-testid="debug-preview-view" />
                 <Tab value="debug" label={t('debug')} data-testid="debug-view-debug" />
                 <Tab value="test" label={t('test')} data-testid="debug-view-tests" />
                 <Tab value="discuss" label={t('discuss')} data-testid="debug-view-collaboration" />
@@ -301,6 +303,8 @@ function ProjectPageView() {
           <Suspense>
             {!file ? (
               <DebugEmptyView />
+            ) : currentTab === 'preview' ? (
+              <PreviewView projectId={projectId} gitRef={gitRef} assistant={file} />
             ) : currentTab === 'debug' ? (
               <DebugView projectId={projectId} gitRef={gitRef} assistant={file} setCurrentTab={setCurrentTab} />
             ) : currentTab === 'test' ? (